IndexedDB/idbobjectstore_get6.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBObjectStore.get() - delete ObjectStore before calling get() </title>
<link rel="author" title="YuichiNukiyama" href="https://github.com/YuichiNukiyama">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>
    var db,
      t = async_test(),
      record = { key: 1, property: "data" };

    var open_rq = createdb(t);
    open_rq.onupgradeneeded = function(e) {
        db = e.target.result;
        db.createObjectStore("store", { keyPath: "key" })
    }

    open_rq.onsuccess = function(e) {
        assert_throws("InvalidStateError", function () {
            var rq = db.transaction("store")
                     .objectStore("store")
            db.deleteObjectStore("store");
            rq.get(1);
        }, "Calling get() should throw InvalidStateError after delete ObjectStore.");
            t.done();
    }
</script>

<div id="log"></div>
